requirements.txt
google-cloud-bigquery==3.11.3
requests==2.28.2